Feature/allocation bi

+++ database/migrations/2025_07_15_163403_create_sales_return_items_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('sales_return_items')) {
+            return;
+        }
+
         Schema::create('sales_return_items', function (Blueprint $table) {
             $table->id();
             $table->foreignId('sales_return_id')->constrained()->onDelete('cascade');
@@ -26,7 +30,7 @@ return new class extends Migration
         Schema::table('sales_return_items', function (Blueprint $table) {           
              $table->foreign('product_id')
                 ->references('id')    // referenced column
-                ->on('supply_profiles')      // referenced table
+                ->on('products')      // referenced table
                 ->onDelete('cascade'); // cascade delete
      
         });
